dr_doctor_scraper/requirements.txt
playwright
beautifulsoup4
pydantic
pymongo
python-dotenv
loguru
pandas
